refactor provider output helpers
